Add doc & Refactor
