+++ app/src/main/java/com/github/factotum_sdp/factotum/models/Bag.kt
@@ -1,5 +1,8 @@
 package com.github.factotum_sdp.factotum.models
 import kotlinx.serialization.Serializable
 
+/**
+ * The Bag data model
+ */
 @Serializable
 class Bag(private val packs: List<Pack>): List<Pack> by packs
