+++ app/src/main/java/com/github/factotum_sdp/factotum/repositories/RoadBookRepository.kt
@@ -24,12 +24,6 @@ import java.util.Locale
  * In charge of being the unique place where to fetch and push RoadBook Data, i.e
  * a DRecordList : List<DestinationRecord>.
  *
- * It resolves the conflict between a "remoteSource" and a "localSource" of data, and enable caching
- * for both remote/network requests and local requests.
- *
- * If remoteSource is not online, we rely uniquely on the localSource, and don't use the first variable cache.
- * The first variable cache is used uniquely to avoid remoteSource update.
- * However, a second variable cache is used before fetching or update the localSource.
  */
 class RoadBookRepository(remoteSource: DatabaseReference, username: String,
                          localSource: DataStore<DRecordList>): BackUpRepository<DRecordList>(remoteSource, username, localSource) {
