+++ app/src/main/java/com/github/factotum_sdp/factotum/serializers/BagSerializer.kt
@@ -1,7 +1,5 @@
 package com.github.factotum_sdp.factotum.serializers
 
-import com.github.factotum_sdp.factotum.models.Pack
-import com.github.factotum_sdp.factotum.ui.roadbook.DRecordList
 import kotlinx.serialization.SerializationException
 import androidx.datastore.core.Serializer
 import com.github.factotum_sdp.factotum.models.Bag
